+++ internal/model/operate_log.go
@@ -5,8 +5,8 @@ import (
 )
 
 type OperateLog struct {
-	ServerID      int64       `json:"server_id"`
-	EnvironmentID int64       `json:"environment_id"`
+	ServerID      uint32      `json:"server_id"`
+	EnvironmentID uint32      `json:"environment_id"`
 	Config        string      `json:"config"`
 	Compare       string      `json:"compare"`
 	Server        Server      `json:"server"`
@@ -17,7 +17,7 @@ func (ol *OperateLog) Create(ctx kratos.Context) error {
 	return ctx.DB().Create(ol).Error
 }
 
-func (ol *OperateLog) Page(ctx kratos.Context, options PageOptions) ([]*OperateLog, int64, error) {
+func (ol *OperateLog) Page(ctx kratos.Context, options PageOptions) ([]*OperateLog, uint32, error) {
 	var list []*OperateLog
 	total := int64(0)
 
@@ -26,14 +26,14 @@ func (ol *OperateLog) Page(ctx kratos.Context, options PageOptions) ([]*OperateL
 		db = db.Scopes(options.Scopes)
 	}
 	if err := db.Count(&total).Error; err != nil {
-		return nil, total, err
+		return nil, uint32(total), err
 	}
 
 	db = db.Offset(int((options.Page - 1) * options.PageSize)).Limit(int(options.PageSize))
 
-	return list, total, db.Find(&list).Error
+	return list, uint32(total), db.Find(&list).Error
 }
 
-func (ol *OperateLog) OneById(ctx kratos.Context, id int64) error {
+func (ol *OperateLog) OneById(ctx kratos.Context, id uint32) error {
 	return ctx.DB().First(ol, "id = ?", id).Error
 }
