+++ internal/model/server.go
@@ -15,7 +15,7 @@ func (s *Server) Create(ctx kratos.Context) error {
 }
 
 // OneByID 通过关键词查找指定资源
-func (s *Server) OneByID(ctx kratos.Context, id int64) error {
+func (s *Server) OneByID(ctx kratos.Context, id uint32) error {
 	return ctx.DB().First(s, "id = ?", id).Error
 }
 
@@ -24,7 +24,7 @@ func (s *Server) OneByKeyword(ctx kratos.Context, keyword string) error {
 }
 
 // Page 查询分页资源
-func (s *Server) Page(ctx kratos.Context, options *PageOptions) ([]*Server, int64, error) {
+func (s *Server) Page(ctx kratos.Context, options *PageOptions) ([]*Server, uint32, error) {
 	var list []*Server
 	total := int64(0)
 
@@ -33,12 +33,12 @@ func (s *Server) Page(ctx kratos.Context, options *PageOptions) ([]*Server, int6
 		db = db.Scopes(options.Scopes)
 	}
 	if err := db.Count(&total).Error; err != nil {
-		return nil, total, err
+		return nil, uint32(total), err
 	}
 
 	db = db.Offset(int((options.Page - 1) * options.PageSize)).Limit(int(options.PageSize))
 
-	return list, total, db.Find(&list).Error
+	return list, uint32(total), db.Find(&list).Error
 }
 
 // All 查询全部资源
@@ -59,6 +59,6 @@ func (s *Server) Update(ctx kratos.Context) error {
 }
 
 // DeleteByID 删除指定id的资源
-func (s *Server) DeleteByID(ctx kratos.Context, id int64) error {
+func (s *Server) DeleteByID(ctx kratos.Context, id uint32) error {
 	return ctx.DB().Model(s).Delete(s, "id = ?", id).Error
 }
