+++ internal/model/template.go
@@ -7,7 +7,7 @@ import (
 
 type Template struct {
 	BaseModel
-	ServerID    int64   `json:"server_id"`
+	ServerID    uint32  `json:"server_id"`
 	Content     string  `json:"content"`
 	Version     string  `json:"version"`
 	IsUse       bool    `json:"is_use"`
@@ -25,16 +25,16 @@ func (t *Template) Create(ctx kratos.Context) error {
 	})
 }
 
-func (t *Template) OneById(ctx kratos.Context, id int64) error {
+func (t *Template) OneById(ctx kratos.Context, id uint32) error {
 	return ctx.DB().First(t, "id = ?", id).Error
 }
 
-func (t *Template) Current(ctx kratos.Context, srvId int64) error {
+func (t *Template) Current(ctx kratos.Context, srvId uint32) error {
 	return ctx.DB().First(t, "is_use = true and server_id = ?", srvId).Error
 }
 
 // Page 查询分页资源
-func (t *Template) Page(ctx kratos.Context, options *PageOptions) ([]*Template, int64, error) {
+func (t *Template) Page(ctx kratos.Context, options *PageOptions) ([]*Template, uint32, error) {
 	var list []*Template
 	total := int64(0)
 
@@ -43,15 +43,15 @@ func (t *Template) Page(ctx kratos.Context, options *PageOptions) ([]*Template, 
 		db = db.Scopes(options.Scopes)
 	}
 	if err := db.Count(&total).Error; err != nil {
-		return nil, total, err
+		return nil, uint32(total), err
 	}
 
 	db = db.Offset(int((options.Page - 1) * options.PageSize)).Limit(int(options.PageSize))
 
-	return list, total, db.Find(&list).Error
+	return list, uint32(total), db.Find(&list).Error
 }
 
-func (t *Template) UseVersionByID(ctx kratos.Context, srvId, id int64) error {
+func (t *Template) UseVersionByID(ctx kratos.Context, srvId, id uint32) error {
 	return ctx.DB().Transaction(func(tx *gorm.DB) error {
 		if err := tx.Where("server_id=? and id=?", srvId, t.ID).Update("is_use", true).Error; err != nil {
 			return err
